Move bigquery queries
