+++ ingestion/src/metadata/ingestion/source/database/bigquery/usage.py
@@ -11,11 +11,11 @@
 """
 Handle big query usage extraction
 """
+from metadata.ingestion.source.database.bigquery.queries import BIGQUERY_STATEMENT
 from metadata.ingestion.source.database.bigquery.query_parser import (
     BigqueryQueryParserSource,
 )
 from metadata.ingestion.source.database.usage_source import UsageSource
-from metadata.utils.sql_queries import BIGQUERY_STATEMENT
 
 
 class BigqueryUsageSource(BigqueryQueryParserSource, UsageSource):
